CryptoBot/requirements.txt
ccxt==4.5.10
python-dotenv==1.1.1
colorama==0.4.6

